+++ CHANGELOG.md
@@ -9,3 +9,23 @@
 ## 0.0.3
 
 * Bug Fixed
+
+## 0.0.4
+
+* Bug Fixed
+
+## 0.0.5
+
+* Bug Fixed
+
+## 0.0.6
+
+* Android FileProvider error fix
+
+## 0.0.7
+
+* Dependencies error fixed
+
+## 0.1.0
+
+* Dependencies to stable channel Flutter